Lint/clean display and move highlighting code.
